Removed pointless cast, as Jamie points out. Because the whole back half of this expression is evaluated at compile time and turns into a long, and then long * int = long.
What wouldn't work is something like:
   longValue = intMethod1() * intMethod2() * LONG_CONSTANT;
if the return values from the two methods overflowed when multiplied.
In that case you'd want to cast the first result to a long
to get things into longspace asap.

+++ src/java/com/threerings/msoy/money/server/MoneyLogic.java
@@ -1339,5 +1339,5 @@ public class MoneyLogic
     protected static final Object LIST_ITEM_KEY = new Object();
 
     /** We don't service bling cashouts at any lower frequency than this. */
-    protected static final long CASHOUT_FREQUENCY = ((long)BlingInfo.CASHOUT_DAYS)*24*60*60*1000L;
+    protected static final long CASHOUT_FREQUENCY = BlingInfo.CASHOUT_DAYS * 24 * 60 * 60 * 1000L;
 }
